+++ web/routes/routes.ts
@@ -55,6 +55,7 @@ export const BACKEND_ROUTES = {
     bulk: `${PRIVATE}/current_account/bulk`,
     calculate: `${PRIVATE}/current_account/calculate`,
     liquidate: `${PRIVATE}/current_account/liquidate`,
+    totals: `${PRIVATE}/current_account/totals`,
   },
   results: {
     base: `${PRIVATE}/results`,